fix(jco): fix invalid types for jco api
