+++ packages/jco/tsconfig.json
@@ -1,8 +1,8 @@
 {
-    "include": ["src/**/*"],
-    "exclude": ["node_modules", "test", "src/cmd/**", "src/*.d.ts", "src/**/*.d.ts"],
+    "include": ["src"],
     "compilerOptions": {
         "outDir": "types",
+        "rootDir": "src",
         "module": "nodenext",
         "moduleResolution": "nodenext",
         "allowJs": true,
