+++ packages/jco/types/cmd/opt.d.ts
@@ -5,15 +5,12 @@ export function opt(componentPath: any, opts: any, program: any): Promise<void>;
  * @param {{ quiet: boolean, asyncify?: boolean, optArgs?: string[], noVerify?: boolean }} opts?
  * @returns {Promise<{ component: Uint8Array, compressionInfo: { beforeBytes: number, afterBytes: number }[] >}
  */
-export function optimizeComponent(
-    componentBytes: Uint8Array,
-    opts: {
-        quiet: boolean;
-        asyncify?: boolean;
-        optArgs?: string[];
-        noVerify?: boolean;
-    },
-): Promise<{
+export function optimizeComponent(componentBytes: Uint8Array, opts: {
+    quiet: boolean;
+    asyncify?: boolean;
+    optArgs?: string[];
+    noVerify?: boolean;
+}): Promise<{
     component: Uint8Array;
     compressionInfo: {
         beforeBytes: number;
